feat(nc-gui): add survey form pages
